MCP/极客时间学习/code/从0到1，基于MCP快速搭建RAG“医疗健康“指北”/rag-server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.10.0",
    "mcp[cli]>=1.6.0",
    "openai>=1.75.0",
]
